+++ src/libk/core/elf.h
@@ -48,7 +48,7 @@ typedef struct
 {
     void (*prints)(const char *str);
     void (*printc)(char c);
-    void (*setcolor)(uint8_t fg, uint8_t bg);
+    void (*setcolor)(uint32_t fg, uint32_t bg);
     void (*log_internal)(const char *file, int line, const char *fmt, int level, int visibility, ...);
     void *(*kmalloc)(uint64_t size);
     void (*kfree)(void *ptr);
